% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{calculate_ARID}
\alias{calculate_ARID}
\title{Calculate ARID Index for Drought Stress}
\usage{
calculate_ARID(ETo, PAW)
}
\arguments{
\item{ETo}{A numeric value representing the reference evapotranspiration.}

\item{PAW}{A numeric value representing the plant available water.}
}
\value{
A numeric value representing the ARID index for drought stress.
}
\description{
This function calculates the ARID index based on reference evapotranspiration (ETo) and plant available water (PAW).
}
\examples{
calculate_ARID(5, 100)
}
