% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{fHeat}
\alias{fHeat}
\title{Impact of Heat Stress on Biomass Growth}
\usage{
fHeat(Tmax, Theat, Textreme)
}
\arguments{
\item{Tmax}{A numeric value representing the daily maximum temperature.}

\item{Theat}{A numeric value representing the heat stress threshold temperature.}

\item{Textreme}{A numeric value representing the extreme temperature threshold.}
}
\value{
A numeric value representing the impact of heat stress on biomass growth.
}
\description{
This function calculates the impact of daily maximum temperature (Tmax) on biomass growth due to heat stress.
If Tmax is less than or equal to Theat, it returns 1.
If Tmax is between Theat and Textreme, it returns a value between 1 and 0 based on the ratio.
If Tmax is greater than Textreme, it returns 0.
}
\examples{
fHeat(35, 30, 40)
}
