% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{fSolar_water}
\alias{fSolar_water}
\title{Calculate Fraction of Solar Radiation Intercepted with Drought Stress}
\usage{
fSolar_water(fWater_value)
}
\arguments{
\item{fWater_value}{A numeric value representing the impact of drought stress on RUE.}
}
\value{
A numeric value representing the fraction of solar radiation intercepted with drought stress.
}
\description{
This function calculates the fraction of solar radiation intercepted based on drought stress.
}
\examples{
fSolar_water(0.05)
}
