% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{fTemp}
\alias{fTemp}
\title{Impact of Temperature on Biomass Growth}
\usage{
fTemp(T, Tbase, Topt)
}
\arguments{
\item{T}{A numeric value representing the daily mean temperature.}

\item{Tbase}{A numeric value representing the base temperature.}

\item{Topt}{A numeric value representing the optimal temperature.}
}
\value{
A numeric value representing the impact of temperature on biomass growth.
}
\description{
This function calculates the impact of daily mean temperature (T) on biomass growth.
If T is less than Tbase, it returns 0.
If T is between Tbase and Topt, it returns a value between 0 and 1 based on the ratio.
If T is greater than or equal to Topt, it returns 1.
}
\examples{
fTemp(25, 10, 30)
}
