% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{fSolar}
\alias{fSolar}
\title{Calculate Fraction of Solar Radiation Intercepted by Crop Canopy During Growth and Senescence Phases}
\usage{
fSolar(TT, I50A, I50B, Tsum, phase)
}
\arguments{
\item{TT}{Thermal time.}

\item{I50A}{Thermal time at which 50% of maximum interception occurs during growth phase.}

\item{I50B}{Thermal time at which 50% of maximum interception occurs during senescence phase.}

\item{Tsum}{Total thermal time for crop growth cycle.}

\item{phase}{Growth phase ("growth" or "senescence").}
}
\value{
Fraction of solar radiation intercepted by crop canopy during growth and senescence phases.
}
\description{
This function calculates the fraction of solar radiation intercepted by a crop canopy during growth and senescence phases.
}
\examples{
fSolar(500, 100, 200, 1000, "growth")
}
