
\name{fSolar_water}
\alias{fSolar_water}
\title{Calculate Fraction of Solar Radiation Intercepted with Drought Stress}
\description{
  This function calculates the fraction of solar radiation intercepted based on drought stress.
}
\usage{
fSolar_water(fWater_value)
}
\arguments{
  \item{fWater_value}{Impact of drought stress on RUE.}
}
\value{
  Fraction of solar radiation intercepted.
}
\examples{
  fSolar_water(0.5)
}
