% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{fWater}
\alias{fWater}
\title{Impact of Drought Stress on RUE}
\usage{
fWater(ARID, Swater)
}
\arguments{
\item{ARID}{A numeric value representing the ARID index for drought stress.}

\item{Swater}{A numeric value representing the sensitivity to water stress.}
}
\value{
A numeric value representing the impact of drought stress on RUE.
}
\description{
This function calculates the impact of drought stress on RUE based on the ARID index and sensitivity (Swater).
}
\examples{
fWater(0.5, 0.01)
}
